Refactor work server
